your message

+++ saudi-pro/saudi_project/admin/delete.php
@@ -1,9 +1,11 @@
 <?php
 include("../config.php");
 
-$id = $_GET['id'];
+$id = isset($_GET['id']) ? (int)$_GET['id'] : 0;
 
-$conn->query("DELETE FROM places WHERE id=$id");
+if($id > 0){
+    $conn->query("DELETE FROM places WHERE id=$id");
+}
 
 header("Location: dashboard.php?msg=deleted");
 exit();
